configure with vite
